+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Teammate H2H point to fantasy league stats
 - Race range selector for prediction league classification
 - ChampionshipMixin and RaceRangeSelectorMixin
+### Changed
+- Race range selector html to template to include
+- Race range selector javascript to its own file
 
 ## [2.16.9] - 2025-03-10
 ### Added